check salary now working

+++ src/main/java/Manager.java
@@ -1,8 +1,8 @@
-public class Manager extends Employee{
+public class Manager extends Employee {
 
     private String deptName;
 
-    public Manager(String name, String deptName, String nationalInsurance , int salary){
+    public Manager(String name, String deptName, String nationalInsurance , double salary){
         super(name, nationalInsurance, salary);
         this.deptName = deptName;
